requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
